requirements.txt
antiorm==1.2.1
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
db==0.1.1
db-sqlite3==0.0.1
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
parameterized==0.8.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
urlparser==0.1.2
